fix(data-table): update to v7
